feat: 사람은 언어를 가진다.
