+++ README.md
@@ -7,3 +7,4 @@
 - [x] 사람은 스킬을 가진다.
 - [x] 사람의 정보는 introduce 함수를 통해 초기화된다.
 - [x] 사람 객체는 불변성을 유지해야한다.
+- [x] 사람은 언어를 가진다.
