+++ src/main/kotlin/dsl/Language.kt
@@ -1,0 +1,3 @@
+package dsl
+
+data class Language(val name: String, val level: Int)
